+++ presentation/src/main/java/com/agileburo/anytype/presentation/linking/LinkToObjectViewModelFactory.kt
@@ -3,13 +3,15 @@ package com.agileburo.anytype.presentation.linking
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.agileburo.anytype.domain.block.interactor.CreateLinkToObject
+import com.agileburo.anytype.domain.config.GetConfig
 import com.agileburo.anytype.domain.misc.UrlBuilder
 import com.agileburo.anytype.domain.page.navigation.GetPageInfoWithLinks
 
 class LinkToObjectViewModelFactory(
     private val urlBuilder: UrlBuilder,
     private val getPageInfoWithLinks: GetPageInfoWithLinks,
-    private val createLinkToObject: CreateLinkToObject
+    private val createLinkToObject: CreateLinkToObject,
+    private val getConfig: GetConfig
 ) : ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
@@ -17,7 +19,8 @@ class LinkToObjectViewModelFactory(
         return LinkToObjectViewModel(
             urlBuilder = urlBuilder,
             getPageInfoWithLinks = getPageInfoWithLinks,
-            createLinkToObject = createLinkToObject
+            createLinkToObject = createLinkToObject,
+            getConfig = getConfig
         ) as T
     }
 }